creating roles to menu and restaurant

+++ src/main/java/br/com/fiap/services/MenuServiceImpl.java
@@ -1,21 +1,26 @@
 package br.com.fiap.services;
 
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
+
 import br.com.fiap.dto.MenuDTO;
 import br.com.fiap.dto.MenuResponseDTO;
 import br.com.fiap.dto.PaginatedResponseDTO;
 import br.com.fiap.exceptions.ResourceNotFoundException;
+import br.com.fiap.exceptions.UnauthorizedException;
 import br.com.fiap.interfaces.repositories.MenuRepository;
 import br.com.fiap.interfaces.services.MenuService;
 import br.com.fiap.interfaces.services.RestaurantService;
+import br.com.fiap.model.AuthenticatedUser;
 import br.com.fiap.model.Menu;
 import br.com.fiap.model.Restaurant;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class MenuServiceImpl implements MenuService {
@@ -28,14 +33,29 @@ public class MenuServiceImpl implements MenuService {
 
 
 	@Override
-	public void save(MenuDTO menuDTO, Long codigoRestaurante) {
-		Restaurant restaurant = restaurantService.getRestaurant(codigoRestaurante);
+	public void save(MenuDTO menuDTO, Long restaurantId) {
+		Restaurant restaurant = restaurantService.getRestaurant(restaurantId);
+		this.save(restaurant, menuDTO);
+	}
+	
+	@Override
+	public void save(MenuDTO menuDTO) {
+		AuthenticatedUser userAuth = (AuthenticatedUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		if(Objects.isNull(userAuth.getRestaurantId())){
+			throw new UnauthorizedException("Restaurante não cadastrado para esse acesso. Favor renove suas credenciais e tente novamente.");
+		}
+		
+		Restaurant restaurant = restaurantService.getRestaurant(userAuth.getRestaurantId());
+		this.save(restaurant, menuDTO);
+	}
+
+	private void save(Restaurant restaurant, MenuDTO menuDTO) {
 		if (restaurant != null) {
 			Menu menu = new Menu(menuDTO, restaurant);
 			menuRepository.save(menu);
 		}
 	}
-
+	
 	@Override
 	public MenuResponseDTO getMenuById(Long id) {
 		var menu = menuRepository.findById(id)
@@ -52,9 +72,12 @@ public class MenuServiceImpl implements MenuService {
 
 	@Override
 	public void update(MenuDTO menuDTO, Long id) {
+		
 		Menu menu = menuRepository.findById(id)
 				.orElseThrow(() -> new ResourceNotFoundException("Cardápio não encontrado com o id: " + id));
 
+		 checkPermissionMenuUpdateDelete(menu);
+		
 		menu.setName(menuDTO.name());
 		menu.setDescription(menuDTO.description());
 		menu.setAvailability(menuDTO.availability());
@@ -63,18 +86,42 @@ public class MenuServiceImpl implements MenuService {
 
 		menuRepository.save(menu);
 	}
+	
+	private void checkPermissionMenuUpdateDelete(Menu menu) {
+		AuthenticatedUser userAuth = (AuthenticatedUser) SecurityContextHolder.getContext().getAuthentication()
+				.getPrincipal();
+		
+		if (!userAuth.hasRoleAdmin() && !menu.getRestaurant().getRestaurantOwner().equals(userAuth.getId())) {
+			throw new UnauthorizedException("Esse menu não pertence ao usuário logado.");
+		}
+	}
 
 	@Override
 	public void delete(Long id) {
 		Menu menu = menuRepository.findById(id)
-				.orElseThrow(
-						() -> new ResourceNotFoundException("Cardápio não encontrado com o id: " + id + "."));
+				.orElseThrow(() -> new ResourceNotFoundException("Cardápio não encontrado com o id: " + id + "."));
+
+		checkPermissionMenuUpdateDelete(menu);
+		
 		menuRepository.delete(menu);
 	}
 
 	@Override
 	public PaginatedResponseDTO<MenuResponseDTO> getAllMenu(Pageable pageable) {
+		AuthenticatedUser userAuth = (AuthenticatedUser) SecurityContextHolder.getContext().getAuthentication()
+				.getPrincipal();
+		Page<Menu> menuPage = menuRepository.findAllByRestaurant(userAuth.getRestaurantId(), pageable);
+
+		return getPage(menuPage);
+	}
+	
+	@Override
+	public PaginatedResponseDTO<MenuResponseDTO> getAllMenu(Pageable pageable, Long restaurantId) {
 		Page<Menu> menuPage = menuRepository.findAll(pageable);
+		return getPage(menuPage);
+	}
+
+	private PaginatedResponseDTO<MenuResponseDTO> getPage(Page<Menu> menuPage) {
 		List<MenuResponseDTO> menuResponseDTOS = menuPage.getContent().stream()
 				.map(menu -> new MenuResponseDTO(menu.getName(), menu.getDescription(), menu.getAvailability(),
 						menu.getPrice(), menu.getPhoto()))
@@ -87,4 +134,5 @@ public class MenuServiceImpl implements MenuService {
 				menuPage.getSize()
 		);
 	}
+
 }
